Xcode and a lot of small changes.

+++ include/result_pair.h
@@ -37,8 +37,8 @@ struct __result_pair
 };
 
 result_pair_t result_pair_create(char *seq1, char *seq2, double value);
-int result_pair_hash_table_destroy(void *data);
-int result_pair_destroy(void *data);
+size_t result_pair_hash_table_destroy(void *data);
+size_t result_pair_destroy(void *data);
 void result_pair_print(char *id, void *data);
 void* result_pair_clone(void *data);
 DISTMAT *convert_to_distmat(hash_table_t result_table);
